requirements/base.txt
# Core Django
Django>=4.2.0,<5.0.0
django-environ>=0.11.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9  # For PostgreSQL
# SQLite comes with Python

# API
djangorestframework>=3.14.0
django-filter>=23.3

# Authentication
djangorestframework-simplejwt>=5.3.0

# Background Tasks
celery>=5.3.0
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0

# Development Tools
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0 